Change selected panel visual
This change alters the visual feedback for a selected panel. Prior to this change the visual feedback for a selected panel was a purple 1px border. This border went around the entire panel. This change updates the visual to highlight the panel toolbar instead of a border.

My opinion is that the toolbar highlight is less disruptive and maintains a cleaner appearence. The panel border would often result in a 1px boundary around some parts of the panel (possibly due to mosaic interactions) which has long been a visual annoyance for me.